--- pkg/utils/testUtils/doc.go
@@ -1,2 +1,0 @@
-// Package testutils contains some methods useful to mock components while testing
-package testutils
