+++ pkg/utils/testutil/dns.go
@@ -1,4 +1,4 @@
-package testutils
+package testutil
 
 import (
 	"net"
