+++ pkg/utils/testutil/doc.go
@@ -1,0 +1,2 @@
+// Package testutil contains some methods useful to mock components while testing
+package testutil
