+++ pkg/utils/testutil/kubernetes.go
@@ -1,4 +1,4 @@
-package testutils
+package testutil
 
 import (
 	corev1 "k8s.io/api/core/v1"
